organized the states into labeled folders
